tests passed

+++ Back end/Polaris/Models/Banking/Transaction.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Models.Banking
 {
     public class Transaction : AmountedEntity<string, string>
@@ -5,7 +7,7 @@ namespace Models.Banking
         public override string Id { get; set; }
         public override string Source { get; set; }
         public override string Target { get; set; }
-        public override double Amount { get; set; }
+        public override Int64 Amount { get; set; }
         public string Timestamp { get; set; }
         public int TrackingId { get; set; }
         public string Type { get; set; }
